appended aka data to /players/{battletag} endpoint
