+++ W3ChampionsStatisticService/Startup.cs
@@ -67,6 +67,7 @@ namespace W3ChampionsStatisticService
             services.AddSpecialBsonRegistrations();
 
             services.AddSingleton<TrackingService>();
+            services.AddSingleton<PlayerAkaProvider>();
 
             services.AddTransient<IMatchEventRepository, MatchEventRepository>();
             services.AddTransient<IVersionRepository, VersionRepository>();
